Add files via upload
Coupling distributes shocks and vibrations to the load, resulting in poor performance. In this study the flexibility of jaw and spider coupling is investigated by using two flexible couplings with different rigidity. Moreover, several tuning methods are applied to design a speed controller, resulting in different solutions accordingly. Each solution can be applied to a specific application based on its constraints and objectives. Criteria such as rising time, overshoot and settling time are used as the basis of comparison.